Implemented BookListsFragment & ListAdapter
