+++ .idea/misc.xml
@@ -12,6 +12,7 @@
         <entry key="app/src/main/res/layout/fragment_search.xml" value="0.155893536121673" />
         <entry key="app/src/main/res/layout/fragment_settings.xml" value="0.351434326171875" />
         <entry key="app/src/main/res/layout/fragment_toolbar.xml" value="0.2503402709960938" />
+        <entry key="app/src/main/res/layout/list_fragment_book.xml" value="0.34791666666666665" />
         <entry key="app/src/main/res/layout/search_result_book.xml" value="0.36936936936936937" />
         <entry key="app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1" />
       </map>
