+++ app/src/main/java/edu/temple/bookrecommendationsystem/BookListsFragment.kt
@@ -1,11 +1,12 @@
 package edu.temple.bookrecommendationsystem
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 
 /*
 A fragment class that composes the Lists tab in the application and contains two BookListFragments -
@@ -27,6 +28,39 @@ class BookListsFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        Log.d("Want to read list: ", Application.Singleton.wantToRead[0].title)
+
+        val wantRecyclerView = requireView().findViewById<RecyclerView>(R.id.want_recycler_view)
+        wantRecyclerView.layoutManager = LinearLayoutManager(requireContext(),
+            LinearLayoutManager.HORIZONTAL, false)
+        wantRecyclerView.adapter = ListAdapter(Application.Singleton.dummyBooks) {
+            val fragment = BookDetailsFragment()
+            val bundle = Bundle()
+            bundle.putString("title", it.title)
+            bundle.putString("author", it.author)
+            bundle.putInt("cover", it.coverURL)
+            fragment.arguments = bundle
+            parentFragmentManager.beginTransaction()
+                .replace(R.id.container1, fragment)
+                .addToBackStack(null)
+                .commit()
+        }
+
+        val prevRecyclerView = requireView().findViewById<RecyclerView>(R.id.prev_recycler_view)
+        prevRecyclerView.layoutManager = LinearLayoutManager(requireContext(),
+            LinearLayoutManager.HORIZONTAL, false)
+        prevRecyclerView.adapter = ListAdapter(Application.Singleton.dummyBooks) {
+            val fragment = BookDetailsFragment()
+            val bundle = Bundle()
+            bundle.putString("title", it.title)
+            bundle.putString("author", it.author)
+            bundle.putInt("cover", it.coverURL)
+            fragment.arguments = bundle
+            parentFragmentManager.beginTransaction()
+                .replace(R.id.container1, fragment)
+                .addToBackStack(null)
+                .commit()
+        }
+
     }
+
 }
